requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
spacy>=3.7.0
python-docx>=1.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
# English model (required at runtime):
#   pip install https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
